+++ src/app/users/page.tsx
@@ -1,1 +1,0 @@
-export { default } from "@/domain/user/screens/list.page";