+++ src/domain/user/user.service.ts
@@ -1,5 +1,0 @@
-export const UserService = {
-    hello() {
-        return "User service is active.";
-    },
-};