+++ src/server/trpc/index.ts
@@ -6,10 +6,10 @@
 // --------------------------------------------------------
 
 import { router } from "./router";
-import { userRouter } from "@/domain/user/user.router";
+
 
 export const appRouter = router({
-  user: userRouter,
+
 });
 
 export type AppRouter = typeof appRouter;